Finished most of the database schema creation. Work on reviews still remains

+++ corner-food-market-website/src/main/java/com/psu/ist311/team5/cornerfoodmarketwebsite/data/repository/ItemImageRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemImageRepository extends JpaRepository<ItemImage, Short> {
+public interface ItemImageRepository extends JpaRepository<ItemImage, Integer> {
     @Query(value = "SELECT II1 FROM ItemImage II1 WHERE II1.itemId = ?1 AND II1.sortNumber = 1")
     ItemImage getMainItemImageByItemId(short id);
